requirements.txt
flask
requests
nibabel
numpy
matplotlib
flask-cors
python-dotenv